Zaqar: support configurable backends
This patch updates the Zaqar profile so that we have
support for configuring alternate versions of the messaging
and management backends.

In Pike instack-undercloud started using the swift/sqlalchemy
backends and the intent here is to update the new containers
undercloud to use a similar default (thus letting us drop Mongodb).

Change-Id: Ie6a56b9163950cee2c0341afa0c0ddce665f3704