Do not fail if PCI device is missing
Fixes a problem where SR-IOV VF count configuration will fail if a
physical function is in use by a guest when 'puppet apply' is executed.
This change substitutes warnings for failures and skips complaints if a
PCI device is unavailable.

Note: this patch has the side-effect of allowing the same configuration
data on hosts that may *not* or *ever* have PCI SR-IOV devices on the
hardware. Time will tell how evil this is in practice.

Closes-Bug: #1701284
Change-Id: I71edc135432ab2193741c37ce977dd11172401e6
(cherry picked from commit d71697a419f459b8fa4b748a64ac6580b3a82456)