Add option for innodb_flush_log_at_trx_commit = 2 for Galera only
The innodb_flush_log_at_trx_commit flag changes the timing
of when the log buffer is written to disk for writes.
At its default of 1, transactions are written to disk
and the buffer flushed on a per-transaction basis; but when
set to 2, the flush of the buffer proceeds only once per
second.  This removes the durability guarantee for the
single node.  However the central concept of Galera is
that durability is achieved via the cluster as a whole,
in that transactions are replicated to other nodes before
the commit succeeds (though not necessarily written to disk
unless wsrep_causal_reads is set).  In this model,
data would only be lost of all nodes of the Galera cluster
were killed within one second of each other.  Percona's
blog post at https://www.percona.com/blog/2014/11/17/typical-misconceptions-on-galera-for-mysql/
recommends that the value of 2 should be considered "safe"
for a Galera cluster unless you are in fact worried that
all three nodes will be powered off simultaneously.

The value here is added as an option only, defaulting
to the usual default of "1", flush per transaction.

Change-Id: Id5a30f1daf978e094a74db2d284febbc9ae64bb3