Implements AIDE Intrusion Detection System
Introduces a puppet service to configure AIDE Intrusion
Detection. This service init's the database and copies the
new database to the active naming. It also sets a cron job,
using email is `AideEmail` is populated, otherwise the reports
are sent to /var/log/aide/.

AIDE rules can be supplied as a hash, and should the rules ever
be changed, the service will populate the new rules and re-init
a fresh integrity database.

Related-Blueprint: tripleo-aide-database
Change-Id: Iac2ceb7fc6b610f8920ae6f75faa2885f3edf6eb