Move gnocchi::api resource to run with wsgi setup
Having this run in step 4 causes a refresh (restart) for httpd, which
in turn is problematic for the gnocchi db upgrade command, since when
it runs httpd is not available at that point. This fixes the issue,
since the API configuration is now ran at the same time as the wsgi
bits.

Change-Id: Ie0ab389a4450bb940757e34d1964423911885fa3
(cherry picked from commit ea79aff996dae9e0dcaf756193f3e103e23bab48)