Allow log path transformation in fluentd glue
Logs in a containerized deployment are not in the same location as on
a baremetal deployment. This commit adds the $fluentd_path_transform
paramter to the fluentd glue module. This is a regular expression that
is used to transform log paths.  To use this feature, include in your
hiera configuration something like:

    tripleo::profile::base::logging::fluentd::fluentd_path_transform:
      - /var/log/
      - /var/log/containers/

Change-Id: I585b6877074353b5de62e5efaabfbe62432c473d
Partial-bug: #1716427
(cherry picked from commit 1ff0903a3950bc4adbc8c84b5153df6ca0fb6a3d)