Allow to configure snmpd_config
Expose a new Puppet parameter to snmp profile, ``snmpd_config`` which
is an array definded to undef by default.
It can be used to override all snmpd configuration for advanced
deployments.
If used, all parameters have to be configured included users and
passwords, which should be the same as given to snmpd_password
and snmpd_user. There is no logic that will verify the content
of ``snmpd_config``.

Example of hieradata which configures snmpd_config:
  snmpd_config:
    - 'createUser ro_snmp_user MD5 "secrete"',
    - 'rouser ro_snmp_user'
    - 'proc neutron-server'
    - 'proc nova-api'

Depends-On: I5c322b3f5350261b58ef065bd3ded0369cadbc4c
Change-Id: Ief2518d5e47137215a34e9ae3b35c27c87fa6e08
Closes-Bug: #1720868