Revert "Revert "Set meta container-attribute-target=host attribute""
This reverts commit 1681d3bceb2834e8788cc4456d65a76bcf4e1e55. 

NOTE: This needs to be tested against scenario004-containers before merging.

This is needed because when we run bundles we actually
want to store attributes on a per-node basis and not on a per-bundle
basis. By activating this attribute pacemaker will pass
some extra OCS_RESKEY_CRM_meta attributes that will help us in this
decision.

We can merge this once we have packages for pacemaker and
resource-agents releases that contain the necessary fixes.

Proper pacemaker and resource-agents are now in the repo [1] so
we can merge it and backport it to pike.

[1] https://buildlogs.centos.org/centos/7/cloud/x86_64/openstack-pike/

Closes-Bug: #1713007

Change-Id: Ie968470126833939c19223f04db29556e550673d