Rely on container setup for haproxy's certificate user and group
If we're on a containerized environment it might be that we don't have
the haproxy user and group (the containerized undercloud). So this sets
it up the owner as root and relies on the containerized setup to fix the
permissions instead. This way, we won't need to have the haproxy package
installed in the host.

Change-Id: If9a86e1e6b450fa4335e8c60aca512d7f4ca5894