Make sure rabbitmq is fully up before creating any rabbitmq resources
Right now after creating the rabbitmq pacemaker resource, we have no
guarantee that rabbit will be up. Let's add the same mechanism we use
today with the galera-ready exec resource. This gives us the guarantee
that once the resource has been created it is up and we can actually
create rabbitmq users (some 3rd party plugins do that, not stock
TripleO).

Specifically, we probe that the '{rabbit,' app shows up in the status,
so we can guarantee that rabbit is running before invoking any other
rabbitmqctl commands.

Change-Id: Ib37eb2e591f97de54ee6449817ae8d70c6541753
Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>