Add unit test for tripleo::haproxy::service_endpoints
In order to get a proper unit test suits, we need to change the
separator from the "dot" to "double-semicolons".

The "." is a reserved character in YAML for hashes. In puppet world,
we commonly use "::".

Unit tests don't work if we let the "dot" separator, and apparently it's
an intended outcome with that kind of syntax.

Fact it's working in the deploy process is kind of black magic (see the
tripleo::firewall::service_rules resource for example). And the "dot"
creates something weird, as all other resources are using the standard
double-semicolon.

Change-Id: I78625a6e69f58dfb2bbad83fdd4f798cd3f4c281
Closes-Bug: 1737086