Fix up the rabbitmq-ready check
So the current rabbitmq-ready exec has a few unexpected problems:

1) The notify mechanism is not being called, but after discussion
we're comfortable in calling this all the time, just like we do this
for galera.
2) Calling rabbitmqctl inside a container is problematic because
the mere invocation of the cluster_status command will actually
spawn an epmd process which will take the epmd port and which will
subsequently make the rabbitmq-bundle started by pacemaker fail to
form a cluster.

For this reason (working around the rabbitmqctl issue is potentially
doable once we upgrade to erlang 19.x but not with older versions)
it is vital that this container gets spawned with /bin/epmd nooped
to /bin/true.

We now only proceed after rabbit tells us that it is part of a cluster.
Just checking for rabbit being up is not enough because if the user gets
created before the node joins a cluster, it might not be replicated
(depending on the timing).

Partial-Bug: #1739026

Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>
Co-Authored-By: John Eckersberg <jeckersb@redhat.com>
Change-Id: I54c541d86782665ae0f689428a16edc155f87993
Depends-On: Ie74a13a6c8181948900ea0de8ee9717f76f3ce79