Allow wsrep_max_ws_rows and wsrep_max_ws_size to use defaults
Starting with MariaDB 10.1.17 / 10.0.27, the wsrep_max_ws_rows
setting suddenly becomes meaningful, as with previous releases
it had no effect [1].   At the same time, MariaDB has changed
the default value from 131072 to "0" which means no limit,
so that their providing this functionality didn't break anything.
This breaks it for tripleo because we are hardcoding that
never-functional default which makes for much too small writesets.

On the MySQL-Galera side, it's unclear if
the wsrep_max_ws_rows setting yet has a meaningful effect, as [2] [3] remain
open but then in [4] they advertise that it *does* work, but
again they've changed the default to 0/no limit.

Additionally, the size of a writeset is more usefully limited
by the wsrep_max_ws_size variable, which here is also hardcoded
to an obselete default of 1G, the aforementioned MariaDB versions
have also bumped the default to 2G.

I've looked up the origin of these settings in [5] and see
no useful rationale for these numbers to be present, rather than
allowing database-provided defaults to be used.

In all cases it is definite that the wsrep_max_ws_rows setting
of 131072 has never been used by any of the MariaDB/MySQL/Galera
variants and we should not be setting this.

[1] https://mariadb.com/kb/en/library/galera-cluster-system-variables/#wsrep_max_ws_rows
[2] https://bugs.launchpad.net/codership-mysql/+bug/1373909
[3] https://github.com/codership/mysql-wsrep/issues/257
[4] http://galeracluster.com/2016/08/announcing-galera-cluster-5-5-50-and-5-6-31-with-galera-3-17/
[5] https://review.openstack.org/#/c/177765/

Change-Id: I0d25666fed15347c8351f127d8f564b2037ca065
Closes-bug: #1738777