Fluentd: Transform path by service
Currently, the configuration delivered via
service_config_settings does not provide the
transformation path regarding the logs for
containerized deployments. So logs are not
correctly configured on fluentd on containerized
deployments.

This path added this feature so the logs are
properly located on fluentd configuration.

Added unit test for this issue

Partial-bug: #1715187
Partial-bug: #1716427
Change-Id: Ie7df4b8b94cb0ae38096ab95800f211ef1cd8455