Enable automatic restart of memcached, apache and mongodb
Prior to Newton, HA deployments had automatic restart of non-Openstack
services on controller nodes. From Newton onward, after we switched to
systemd for the vast majority of services, Openstack services are
configured for automatic restart, but memcached apache and mongodb
lack it.

Generate a drop-in file for the mentioned services, so they can
restart automatically in both HA and non-HA deployments.

Change-Id: I8576db193b358cf804e8aaeb0e05b31c635ae5ba
Closes-Bug: #1741263