Fix rabbitmq-ready check for single node HA deployments
The current rabbitmq-ready exec waits for rabbitmq to become clustered
before it allows user creation. Unfortunately this doesn't work when
the deployment contains a single node, because rabbit doesn't trigger
the clustering mode at all.

Set the exec test according to the number of rabbit nodes, in order
to check for cluster state only when necessary.

Closes-Bug: #1741345

Change-Id: I24e5e344b7f657ce5d42a7c7c45be7b5ed5e6445
Co-Authored-By: John Eckersberg <jeckersb@redhat.com>