Haproxy endpoint resource: set defaults for server_names and IPs
It's not trivial to set the server names and IP addresses for dynamic
haproxy endpoints due to the fact that hiera renders the array as a
string. In order to make this work, we try to be clever and get the
server names and IPs from hiera directly as a default value. This should
work as long as the endpoint uses the same name as the service.

Closes-Bug: #1743626
Change-Id: I51eebfeceb5915fb0add8332c50386f6df93dca5