Enable automatic restart of memcached, apache and mongodb
Prior to Newton, HA deployments had automatic restart of non-Openstack
services on controller nodes. From Newton onward, after we switched to
systemd for the vast majority of services, Openstack services are
configured for automatic restart, but memcached apache and mongodb
lack it.

Generate a drop-in file for the mentioned services, so they can
restart automatically in both HA and non-HA deployments.

This change is similar to the original review in master [1], but
it is implemented without puppet-systemd, because all stable versions
upstream (N, O, P) ship with an outdated puppet-systemd and one
impacted version downstream (N) doesn't even ship puppet-systemd.

[1] I8576db193b358cf804e8aaeb0e05b31c635ae5ba

Change-Id: I7658f539db55b7416a9931822cfe2102c3388123
Closes-Bug: #1741263