Enable automatic restart of memcached, apache and mongodb
Prior to Newton, HA deployments had automatic restart of non-Openstack
services on controller nodes. From Newton onward, after we switched to
systemd for the vast majority of services, Openstack services are
configured for automatic restart, but memcached apache and mongodb
lack it.

Generate a drop-in file for the mentioned services, so they can
restart automatically in both HA and non-HA deployments.

This change differs slighly from the review backported to Ocata [1]
because in Newton, there is no tripleo-specific profile to configure
Apache; So we generate the drop-in file in the keystone profile to
guarantee that Apache automatic restart will be enabled on nodes where
OpenStack API are deployed.

[1] I7658f539db55b7416a9931822cfe2102c3388123

Change-Id: Ic6b57ec30c53a077eb83afd5c99f5c59e6ff56d3
Closes-Bug: #1741263
Co-Authored-By: Michele Baldessari <michele@acksyn.org>