Use --live-restore when setting up the docker service
If you rerun an overcloud deploy command or do a minor update, while you
changed any of the options in the tripleo::profile::base::docker puppet
manifest, you are at risk of triggering a restart of the docker service
(e.g. you changed the url in the insecure_registries parameter).

This has a number of drastic consequences:
1) *All* the containers will be shut-down and restarted (this will
effectively bring down all the APIs at the same time in the overcloud
deploy case)
2) Pacemaker can potentially fail to start/stop/monitor the bundles (due
to the docker API not being available) which can have a few possible
outcomes (depending on timing): it could fence a node in case of failure
to stop, it could stop any of the bundles, it could recover from such a
situation.

Effectively now that we run stuff inside containers, restarting docker
is a bit like taking down PID 1 and starting it again.

In order to ameliorate this situation we make sure that docker is
actually started with the '--live-restore' option which was introduced
in docker 1.12 via https://github.com/moby/moby/issues/6851 and avoids
restarting all the containers when the docker daemon gets restarted.

Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>

Change-Id: Ib3ea6de7f235d2a2d53a6576e0876ab171128b34
Closes-Bug: #1747851