Avoid hard-coded settings in Cinder HA containers
Add parameters for controlling the docker container settings used to
create the cinder-volume and cinder-backup pacemaker bundles. The
parameters eliminate the need to hard-code the list of docker volumes
and environment variables, making it possible to control the values
using hiera data.

For backward compatibility, the previous hard-coded values are used
when no parameter inputs are supplied.

Partial-Bug: #1748290
Change-Id: I4ba0d78ad17183b97290b853a6c103e55bc8977c