Make exposing the keystone admin port optional
A recent commit [1] removed the exposure of the keystone admin port for
the public interface. While this removal was intended, it was not ideal
to backport it to stable/newton as some folks were relying on this port
being available.

This commit then adds a flag to expose or hide the keystone admin
endpoint from the public network. It's on by default to keep the
functionality that folks with newton were seeing; but it lets folks opt
out of it if needed (as is the case is some deployments).

[1] I79563f62fd49a4f7654779157ebda3c239d6dd22

Closes-Bug: #1748872
Change-Id: I865aff9d54ebbe6acc44ffca41c951a05da71218