Make sure that ceph-nfs and the VIP start in the right order
Currently there is a race where ceph-nfs is free to be started before
the colocation constraint with the VIP is created. This issue is
exacerbated by the fact that the container being started (via a systemd
unit file managed by pacemaker) never fails in case the daemon
encounters an error, because it has inside it:
/usr/bin/ganesha.nfsd "${GANESHA_OPTIONS[@]}" -L /var/log/ganesha/ganesha.log "${GANESHA_EPOCH}" || return 0

Since the container never fails, docker never exits and so systemd and
pacemaker think everything is fine and it won't get rescheduled on
another node.

We can avoid this issue completely if we can guarantee that
we start the VIP and the ceph-nfs service on the bootstrap node only
(by tagging only that node with the right property) and only later we
add the node property on all the other nodes.

The drawback of this approach is that, barring any failures, ceph-nfs
will always be started on the master node for that service.

Change-Id: I78c16c9fb28211a7a8ec2187ae76b9a072b76ea2
(cherry picked from commit 318e20a828ebf70e4a2644148c9b921d7007b5f0)