Extract local CA if it expired
This adds a conditional that extracts certmonger's local CA if the
certificate doesn't exist or if it has expired already. This adds the
ability for the deployer to fix the undercloud installation with the
undercloud install command itself if expiration of the CA cert happens.

Change-Id: I61577be2434d7321dd462902d386c6911c2c4f57
Closes-Bug: #1753948
(cherry picked from commit a4ae09d169396f63b1952570b2105d695753fa9d)