firewall: don't reload IPtables after cleanup
This patch stops the IPtables reload when doing Neutron rules cleanup.

Full context:
puppetlabs-firewall only manages the current state of iptables
rules and writes out the rules to a file to ensure they are
persisted. We are specifically running the following commands after the
iptables rules to ensure the persisted file does not contain any
ephemeral neutron rules. Neutron assumes the iptables rules are not
persisted so it may cause an issue if the rule is loaded on boot
(or via iptables restart). If an operator needs to reload iptables
for any reason, they may need to manually reload the appropriate
neutron agent to restore these iptables rules.

rhbz#1541528
Related-Bug: #1747960
Change-Id: I1ab3a52306b91baadb70d2210a378417087f1ecf