Pin puppet-systemd
puppet-systemd recently added a breaking change to pull the module
parameter defaults via hiera. This only works in puppet >= 4.10.0 which
we are not using. So we need to pin to an earlier commit since we're
usign 4.8.x.

Change-Id: I11c8372eb20d9ff67bcc91897118061d7f193a54
Closes-Bug: #1755211
(cherry picked from commit bd5599ca59b168ce4785991c4165fa4196b8d78a)