Fixes incorrect ownership of ODL TLS cert/key
Deployments were failing because the owner/group of the TLS generated
certificate and key were set to 'odl'.  This user and group does not
exist in a containerized deployment because the ODL RPM is only
installed in the container.

This patch leaves the owner as root for the files which works because
the files are only used to generate a keystore for ODL (which is owned
by odl), and the cert/key files themselves are never read by ODL.

Closes-Bug: 1757135

Change-Id: Ie5b9e98ea2fc16b820d56272653df4874e81cf68
Signed-off-by: Tim Rozet <trozet@redhat.com>