Removes neutron ownership of certs
Since neutron UID is not static, setting the owners on the certificates
in the host to be 'neutron' will not match the UID for neutron in the
deployed container.  Therefore this patch removes the host neutron
ownership and leaves it as root, so that it can be later modified in the
container to be chowned to neutron.

Partial-Bug: 1759049

Change-Id: I83b14b91d1ee600bd9d5863acba34303921368ce
Signed-off-by: Tim Rozet <trozet@redhat.com>