Include local CA installation outside of HAProxy cert setup
It used to be the case that we would only do the local CA setup if we
would require a certificate for HAProxy. This is no longer the case as
we're trying to do TLS by default in the overcloud. So this instead
moves that setup to the certmonger_user class which should always get
used.

Change-Id: I5b4dc0a60ccc988b1ffbbb04231e6688377dd0eb