haproxy: enable UI on containerized undercloud
TripleO UI endpoint was never enabled in HAproxy because enable_ui is
for instack-undercloud.
It should be tripleo_ui_enabled to follow the pattern with other
services, and be automatically set to True when the service is actually
enabled, which is the case of the containerized undercloud.

For backward compatibility, we'll keep support for enable_ui and use
pick to find which one is defined, otherwise we'll just disable it like
before.

Change-Id: I2f9ef89b6daac2dec4e2af110f5d239a3611f747