Partially revert "Fail more gracefully when passed an empty ip"
This reverts the manifests/haproxy.pp parts of
commit 4c7ca4cbc385fa70438015ebc91b8692c00d46e7.

The reverted commit made it such that you can not set a
hostname as undercloud_public_host in undercloud.conf. This
is required so that the UI can be accessed by hostname
externally as well as to make the Undercloud API accessible
externally. We keep the parts of that change that verify if
the haproxy VIP for pacemaker is a proper IP as it can
happen that a misconfigured network template causes empty
strings and the deploy will fail in a very odd way.

Change-Id: I57eeb2dd26336465627593d633c63b92da42e71d
Closes-Bug: #1763776