Fixes binding type for OpenDaylight Websocket
For OpenDaylight Websocket connections we were not using transparent
binding type with HA Proxy.  This means that HA Proxy was not able to
start on nodes that did not have the VIP because it was unable to bind
to that IP on more than one node.  However, transparent binding works OK
with OpenDaylight Websocket and should be fine to enable so that HA
Proxy is able to start on every controller.

Closes-Bug: 1764514

Change-Id: I89e6115795ece6735e816ab71b5b552b17f7b943
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 70bedeef99d5b2852307158330f78c0d9159532f)